+++ bindings/python/examples/how_tos/alias/create.py
@@ -21,5 +21,5 @@ if 'STRONGHOLD_PASSWORD' not in os.environ:
 wallet.set_stronghold_password(os.environ["STRONGHOLD_PASSWORD"])
 
 # Send transaction.
-transaction = account.prepare_create_alias_output(None, None).send()
+transaction = account.create_alias_output(None, None)
 print(f'Block sent: {os.environ["EXPLORER_URL"]}/block/{transaction.blockId}')