+++ bindings/python/iota_sdk/wallet/account.py
@@ -78,6 +78,12 @@ class Account:
         return AccountMetadata(
             self.meta["alias"], self.meta["coinType"], self.meta["index"])
 
+    def burn(
+            self, burn: Burn, options: Optional[TransactionOptions] = None) -> Transaction:
+        """A generic function that can be used to burn native tokens, nfts, foundries and aliases.
+        """
+        return self.prepare_burn(burn, options).send()
+
     def prepare_burn(
             self, burn: Burn, options: Optional[TransactionOptions] = None) -> PreparedTransaction:
         """A generic `prepare_burn()` function that can be used to prepare the burn of native tokens, nfts, foundries and aliases.
@@ -119,6 +125,12 @@ class Account:
         )
         return PreparedTransaction(self, prepared)
 
+    def consolidate_outputs(
+            self, params: ConsolidationParams) -> Transaction:
+        """Consolidate outputs.
+        """
+        return self.prepare_consolidate_outputs(params).send()
+
     def prepare_consolidate_outputs(
             self, params: ConsolidationParams) -> PreparedTransaction:
         """Consolidate outputs.
@@ -129,6 +141,13 @@ class Account:
             }
         )
         return PreparedTransaction(self, prepared)
+
+    def create_alias_output(self,
+                            params: Optional[CreateAliasOutputParams] = None,
+                            options: Optional[TransactionOptions] = None) -> Transaction:
+        """Create an alias output.
+        """
+        return self.prepare_create_alias_output(params, options).send()
 
     def prepare_create_alias_output(self,
                                     params: Optional[CreateAliasOutputParams] = None,
@@ -272,6 +291,12 @@ class Account:
         )
         return [Transaction.from_dict(tx) for tx in transactions]
 
+    def create_native_token(self, params: CreateNativeTokenParams,
+                            options: Optional[TransactionOptions] = None) -> Transaction:
+        """Create native token.
+        """
+        return self.prepare_create_native_token(params, options).send()
+
     def prepare_create_native_token(self, params: CreateNativeTokenParams,
                                     options: Optional[TransactionOptions] = None) -> PreparedTransaction:
         """Create native token.
@@ -284,6 +309,16 @@ class Account:
         )
         return PreparedCreateTokenTransaction(
             account=self, prepared_transaction_data=prepared)
+
+    def melt_native_token(self,
+                          token_id: HexStr,
+                          melt_amount: int,
+                          options: Optional[TransactionOptions] = None) -> Transaction:
+        """Melt native tokens. This happens with the foundry output which minted them, by increasing it's
+        `melted_tokens` field.
+        """
+        return self.prepare_melt_native_token(
+            token_id, melt_amount, options).send()
 
     def prepare_melt_native_token(self,
                                   token_id: HexStr,
@@ -301,6 +336,13 @@ class Account:
         )
         return PreparedTransaction(self, prepared)
 
+    def mint_native_token(self, token_id: HexStr, mint_amount: int,
+                          options: Optional[TransactionOptions] = None) -> Transaction:
+        """Mint additional native tokens.
+        """
+        return self.prepare_mint_native_token(
+            token_id, mint_amount, options).send()
+
     def prepare_mint_native_token(self, token_id: HexStr, mint_amount: int,
                                   options: Optional[TransactionOptions] = None) -> PreparedTransaction:
         """Mint additional native tokens.
@@ -313,6 +355,12 @@ class Account:
             }
         )
         return PreparedTransaction(self, prepared)
+
+    def mint_nfts(self, params: List[MintNftParams],
+                  options: Optional[TransactionOptions] = None) -> Transaction:
+        """Mint NFTs.
+        """
+        return self.prepare_mint_nfts(params, options).send()
 
     def prepare_mint_nfts(self, params: List[MintNftParams],
                           options: Optional[TransactionOptions] = None) -> PreparedTransaction:
@@ -360,6 +408,12 @@ class Account:
             }
         )
         return PreparedTransaction(self, prepared)
+
+    def send_transaction(
+            self, outputs: List[Output], options: Optional[TransactionOptions] = None) -> Transaction:
+        """Send a transaction.
+        """
+        return self.prepare_transaction(outputs, options).send()
 
     def prepare_transaction(
             self, outputs: List[Output], options: Optional[TransactionOptions] = None) -> PreparedTransaction:
@@ -420,6 +474,12 @@ class Account:
             }
         ))
 
+    def send_native_tokens(
+            self, params: List[SendNativeTokensParams], options: Optional[TransactionOptions] = None) -> Transaction:
+        """Send native tokens.
+        """
+        return self.prepare_send_native_tokens(params, options).send()
+
     def prepare_send_native_tokens(
             self, params: List[SendNativeTokensParams], options: Optional[TransactionOptions] = None) -> PreparedTransaction:
         """Send native tokens.
@@ -431,6 +491,12 @@ class Account:
             }
         )
         return PreparedTransaction(self, prepared)
+
+    def send_nft(self, params: List[SendNftParams],
+                 options: Optional[TransactionOptions] = None) -> Transaction:
+        """Send nft.
+        """
+        return self.prepare_send_nft(params, options).send()
 
     def prepare_send_nft(self, params: List[SendNftParams],
                          options: Optional[TransactionOptions] = None) -> PreparedTransaction: