+++ sdk/src/types/block/output/feature/metadata.rs
@@ -74,6 +74,7 @@ impl core::fmt::Debug for MetadataFeature {
     }
 }
 
+#[cfg(feature = "serde")]
 pub(crate) mod dto {
     use alloc::boxed::Box;