+++ sdk/tests/wallet/wallet_storage.rs
@@ -316,6 +316,41 @@ async fn check_existing_db_4() -> Result<()> {
     tear_down(storage_path)
 }
 
+// Db created with iota-sdk commit 37edd0706ee003a0d17c7da19ba974b17b365cfe (@iota/wallet@2.0.2-alpha.32)
+#[cfg(feature = "stronghold")]
+#[tokio::test]
+async fn check_existing_db_5() -> Result<()> {
+    let storage_path = "check_existing_5_db_test";
+    setup(storage_path)?;
+    // Copy db so the original doesn't get modified
+    copy_folder("./tests/wallet/fixtures/check_existing_5_db_test", storage_path).unwrap();
+
+    let wallet = Wallet::builder().with_storage_path(storage_path).finish().await?;
+
+    // Commented because it wasn't created with encrypt_work_factor 0
+    // wallet.set_stronghold_password("STRONGHOLD_PASSWORD".to_owned()).await?;
+
+    assert_eq!(wallet.get_accounts().await?.len(), 1);
+
+    let client_options = wallet.client_options().await;
+    assert_eq!(client_options.node_manager_builder.nodes.len(), 1);
+
+    let account = wallet.get_account("Alice").await?;
+
+    let addresses = account.addresses().await?;
+    // One public address
+    assert_eq!(addresses.len(), 1);
+    // Wallet was created with mnemonic: "endorse answer radar about source reunion marriage tag sausage weekend frost
+    // daring base attack because joke dream slender leisure group reason prepare broken river"
+    assert_eq!(
+        addresses[0].address().to_string(),
+        "rms1qpllaj0pyveqfkwxmnngz2c488hfdtmfrj3wfkgxtk4gtyrax0jaxzt70zy"
+    );
+    assert!(!addresses[0].internal());
+
+    tear_down(storage_path)
+}
+
 fn copy_folder(src: impl AsRef<Path>, dest: impl AsRef<Path>) -> io::Result<()> {
     fs::create_dir_all(&dest)?;
     for entry in fs::read_dir(src)? {