+++ bindings/python/CHANGELOG.md
@@ -25,6 +25,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `ConflictReason` display implementation with an explanation of the conflict;
+- `Account::{burn(), consolidate_outputs(), create_alias_output(), create_native_token(), melt_native_token(), mint_native_token(), mint_nfts(), send_transaction(), send_native_tokens(), send_nft()}` methods;
+
+## 1.0.1 - 2023-08-23
+
+### Fixed
+
+- Ledger Nano events properly created when preparing transactions;
+- `Account::prepare_output()` when `ReturnStrategy.Gift` is used;
 
 ## 1.0.0 - 2023-07-24